Change edict_s->show_hostile from qboolean to int and add casts
This variable is used as an integer. Defining it as a qbooblean is
just wrong. Add some explicit float -> int casts while at it.

+++ action/g_ai.c
@@ -335,7 +335,7 @@ void FoundTarget (edict_t *self)
                 level.sight_entity->light_level = 128;
         }
 
-        self->show_hostile = level.time + 1;            // wake up other monsters
+        self->show_hostile = (int)level.time + 1;            // wake up other monsters
 
         VectorCopy(self->enemy->s.origin, self->monsterinfo.last_sighting);
         self->monsterinfo.trail_time = level.time;
@@ -487,7 +487,7 @@ qboolean FindTarget (edict_t *self)
 
                 if (r == RANGE_NEAR)
                 {
-                        if (client->show_hostile < level.time && !infront (self, client))
+                        if (client->show_hostile < (int)level.time && !infront (self, client))
                         {
                                 return false;
                         }
@@ -777,7 +777,7 @@ qboolean ai_checkattack (edict_t *self, float dist)
                         }
                         else
                         {
-                                self->show_hostile = level.time + 1;
+                                self->show_hostile = (int)level.time + 1;
                                 return false;
                         }
                 }
@@ -843,7 +843,7 @@ qboolean ai_checkattack (edict_t *self, float dist)
                 }
         }
 
-        self->show_hostile = level.time + 1;            // wake up other monsters
+        self->show_hostile = (int)level.time + 1;            // wake up other monsters
 
 // check knowledge of enemy
         enemy_vis = visible(self, self->enemy);
